etl/requirements.txt
psycopg2==2.9.6
pydantic==2.1.1
python-dotenv==1.0.0
lxml==5.2.1